refact: 신고 코드 이동
